requirements.txt
# Core dependencies for Hybrid AI Development Workflow
watchdog>=3.0.0
pyyaml>=6.0
graphviz>=0.20.0

# Testing dependencies
pytest>=7.0.0

# Web framework and scheduling
schedule>=1.2.0
requests>=2.28.0
flask>=2.2.0

# Code quality tools
flake8>=5.0.0
black>=22.0.0
isort>=5.10.0
mypy>=1.0.0
typer>=0.9.0
rich>=13.0.0
blake3>=0.4.0

# Performance monitoring
psutil>=5.9.0

# Graph processing
networkx>=3.0

# Development and utilities
pathlib2>=2.3.0; python_version < "3.4"
